+++ C++/StringPermutations.cpp
