+++ C++/kadanesalgo.cpp
