+++ C++/linearsearch_incpp.cpp
