+++ C++/linked list.cpp
