+++ C/array.c
