+++ Java/crypto-solve.java
